back to single file using #ifdef _BLUEZ_HCI_
